+++ Changelog.md
@@ -18,6 +18,10 @@ v2.36.0 (not published yet)
 ====================
 
 - UI: "Import VUI metadata" doesn't override Dolby Vision profile depending on selected mode ([#1212](/../../issues/1212))
+- NVEncC: Add support for Dolby Vision metadata handling
+- QSVEncC: Add support for Dolby Vision metadata handling
+- VCEEncC: Add support for Dolby Vision metadata handling
+- VCEEncC: Add Dolby Vision parameters to encoder options
 - Update tools
     - ...
 - Update AviSynth+ plugins